// src/types/ShortGithubInfo.ts
export default interface ShortGithubInfo {
	url: string;
	createdAt: Date;
	updatedAt: Date;
	language: string;
	name: string;
	description: string;
}